Transfer translations and modifications from old repo to new one
- Made lowercase text possible
- Modified HalfWidthFont to add more character variations like é and ü, also adding ' as a possible character on the HalfWidthFont.
- Modified HalfWidthFont to make the lowercase text more readable.
- Added a variable called HypnosisAttackName to separate the Attack (Hypnose) from the Pokémon type (Hypnose-Pokémon).
- Translated Title Screen, Booster Pack and Duel Event Message box graphics.
- Translated a lot of text into Dutch.
------
NOTE:
Increasing the size of the KeyboardData_Player variable crashes the autodeck machines as Bank 6 isn't completely disassembled yet. Reverted the KeyboardData_Player variable to its defaults. Until bank 6 is completely disassembled I'm gonna leave the player keyboard for what it is and continue translating the rest with the save games I've already prepared. I've added a backup of my version of the KeyboardData_Player variable in a .txt file.